pyproject.toml
[project]
name = "melius"
version = "0.1.0"
description = "A headless, local-first AI agent CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click",
    "requests",
    "pydantic",
    "python-dotenv",
    "apscheduler",
    "playwright",
    "python-telegram-bot",
    "rich",
    "aiohttp"
]

[project.scripts]
melius = "melius.main:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
